chore(frontend): Scaffoldの分離
